// src/app/model/requestDTO.ts
export interface RequestDTO{

    tipoDocumentoid?: number;
    nro_documento?: string;
    razon_social?: string;
    nombre_comercial?: string;
    tipoContribuyenteid?: number
    direccion?: string;
    telefono?: string;
    // constructor(){
    // this.id_entidad = 0;
    // this.TipoContribuyente = new TipoContribuyente;
    // this.nro_documento = "";
    // this.razon_social = "";
    // this.nombre_comercial = "";
    // this.direccion = "";
    // this.telefono = "";
    // this.estado = true;
    // }
}